requirements.txt
requests
dnspython
pyyaml

